requirements.txt
appdirs==1.4.4
asgiref==3.8.1
beautifulsoup4==4.13.3
certifi==2025.1.31
charset-normalizer==3.4.1
Django==5.2
djangorestframework==3.16.0
dnspython==2.7.0
dotenv==0.9.9
google-play-scraper==1.2.7
idna==3.10
importlib_metadata==8.6.1
pyee==11.1.1
pymongo==4.13.0
pyppeteer==2.0.0
python-dotenv==1.1.0
requests==2.32.3
soupsieve==2.6
sqlparse==0.5.3
tqdm==4.67.1
typing_extensions==4.13.1
urllib3==1.26.20
websockets==10.4
zipp==3.21.0
